fix(build): cast getMouseScroll() to float in SettingsScreen
MSVC promotes C4244 (int-to-float narrowing) to an error under
/WX. getMouseScroll() returns int; add static_cast<float> at the
assignment site.

+++ game/fighters-legacy/SettingsScreen.cpp
@@ -140,7 +140,7 @@ Screen SettingsScreen::update(IInput& input, IWindow& window) {
                       input.isGamepadButtonJustPressed(0, GamepadButton::DpadLeft);
     const bool right = input.isKeyJustPressed(Key::ArrowRight) || input.isKeyJustPressed(Key::D) ||
                        input.isGamepadButtonJustPressed(0, GamepadButton::DpadRight);
-    const float scroll = input.getMouseScroll();
+    const float scroll = static_cast<float>(input.getMouseScroll());
 
     switch (m_focusedRow) {
     case 0: // Resolution
